+++ unique_toolkit/unique_toolkit/chat/__init__.py
@@ -1,5 +1,9 @@
 from .schemas import ChatMessage as ChatMessage
 from .schemas import ChatMessageRole as ChatMessageRole
+from .schemas import MessageAssessment as MessageAssessment
+from .schemas import MessageAssessmentLabel as MessageAssessmentLabel
+from .schemas import MessageAssessmentStatus as MessageAssessmentStatus
+from .schemas import MessageAssessmentType as MessageAssessmentType
 from .service import ChatService as ChatService
 from .utils import (
     convert_chat_history_to_injectable_string as convert_chat_history_to_injectable_string,