+++ unique_sdk/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.19] - 2025-01-31
+- Add `MessageAssessment` resource
+
 ## [0.9.18] - 2025-01-22
 - Removed `Invalid response body from API` from `retry_dict` as it's our own artificail error.